Move methods of PebbleNotificationHandler into PebbleCommunicatorService
